# server/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Admin System API",
    "version": "1.0.0",
    "description": "后台管理系统 API 文档 - 自动生成"
  },
  "servers": [
    {
      "url": "http://localhost:8080/"
    }
  ],
...
  "paths": {
...
    "/manage/sys/user/getUserList": {
      "post": {
        "tags": [
          "用户管理"
        ],
        "summary": "获取用户列表（分页）",
        "description": "",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "rows": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "userId": {
                                "type": "string",
                                "example": "507f1f77bcf86cd799439011"
                              },
                              "realName": {
                                "type": "string",
                                "example": "张三"
                              },
                              "email": {
                                "type": "string",
                                "example": "zhangsan@example.com"
                              },
                              "avatarUrl": {
                                "type": "string",
                                "example": "https://example.com/avatar.jpg"
                              },
                              "status": {
                                "type": "number",
                                "example": 1
                              },
                              "createdAt": {
                                "type": "string",
                                "example": "2024-01-01T00:00:00.000Z"
                              },
                              "roles": {
                                "type": "array",
                                "items": {
                                  "type": "object",
                                  "properties": {
                                    "id": {
                                      "type": "string",
                                      "example": "507f1f77bcf86cd799439012"
                                    },
                                    "roleName": {
                                      "type": "string",
                                      "example": "管理员"
                                    }
                                  }
                                }
                              }
                            }
                          }
                        },
                        "total": {
                          "type": "number",
                          "example": 100
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "page": {
                    "type": "number",
                    "description": "页码",
                    "example": 1
                  },
                  "rows": {
                    "type": "number",
                    "description": "每页数量",
                    "example": 10
                  },
                  "status": {
                    "type": "number",
                    "description": "状态筛选 1:启用 0:禁用",
                    "example": 1
                  }
                }
              }
            }
          }
        }
      }
    },
    "/manage/sys/user/addUser": {
      "post": {
        "tags": [
          "用户管理"
        ],
        "summary": "添加用户",
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "realName",
                  "email",
                  "pwd"
                ],
                "properties": {
                  "realName": {
                    "type": "string",
                    "description": "真实姓名",
                    "example": "张三"
                  },
                  "email": {
                    "type": "string",
                    "description": "邮箱",
                    "example": "zhangsan@example.com"
                  },
                  "pwd": {
                    "type": "string",
                    "description": "密码",
                    "example": "123456"
                  },
                  "roleId": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "description": "角色ID数组",
                    "example": [
                      "507f1f77bcf86cd799439012"
                    ]
                  }
                }
              }
            }
          }
        }
      }
    },
    "/manage/sys/user/editUser": {
      "post": {
        "tags": [
          "用户管理"
        ],
        "summary": "编辑用户",
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "id"
                ],
                "properties": {
                  "id": {
                    "type": "string",
                    "description": "用户ID",
                    "example": "507f1f77bcf86cd799439011"
                  },
                  "realName": {
                    "type": "string",
                    "description": "真实姓名",
                    "example": "张三"
                  },
                  "email": {
                    "type": "string",
                    "description": "邮箱",
                    "example": "zhangsan@example.com"
                  },
                  "roleId": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "description": "角色ID数组",
                    "example": [
                      "507f1f77bcf86cd799439012"
                    ]
                  }
                }
              }
            }
          }
        }
      }
    },
    "/manage/sys/user/deleteUser": {
      "post": {
        "tags": [
          "用户管理"
        ],
        "summary": "删除用户",
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "userId"
                ],
                "properties": {
                  "userId": {
                    "type": "string",
                    "description": "用户ID",
                    "example": "507f1f77bcf86cd799439011"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/manage/sys/user/changeUserStatus": {
      "post": {
        "tags": [
          "用户管理"
        ],
        "summary": "修改用户状态",
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "userId",
                  "status"
                ],
                "properties": {
                  "userId": {
                    "type": "string",
                    "description": "用户ID",
                    "example": "507f1f77bcf86cd799439011"
                  },
                  "status": {
                    "type": "number",
                    "description": "状态 1:启用 0:禁用",
                    "example": 1,
                    "enum": [
                      0,
                      1
                    ]
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}